+++ lib/ruby-next/core/enumerator/produce.rb
@@ -1,13 +1,21 @@
 # frozen_string_literal: true
 
-RubyNext::Core.patch Enumerator.singleton_class, method: :produce, singleton: Enumerator, version: "2.7" do
+RubyNext::Core.patch Enumerator.singleton_class, method: :produce, singleton: Enumerator,
+  supported: Enumerator.respond_to?(:produce) && (
+    begin
+      Enumerator.produce(a: 1, b: 1) {}
+      false
+    rescue ArgumentError
+      true
+    end
+  ),
+  version: "4.0" do
   <<-'RUBY'
-# Based on https://github.com/zverok/enumerator_generate
-def produce(*rest, &block)
+def produce(*rest, size: Float::INFINITY, &block)
   raise ArgumentError, "wrong number of arguments (given #{rest.size}, expected 0..1)" if rest.size > 1
   raise ArgumentError, "No block given" unless block
 
-  Enumerator.new(Float::INFINITY) do |y|
+  Enumerator.new(size) do |y|
     val = rest.empty? ? yield() : rest.pop
 
     loop do